Add DiscordInteractor, refactor folder structure
The DiscordInteractor handles responses to Interactions, for example
an interaction response (which was automatically created on the client
when a command was issued, with content indicating the "waiting" state)
can be updated when a response is ready.

Model classes are the classes that were created to model the Discord
Application domain objects. All model classes were moved inside a folder
called "models" for findability, now that we are adding classes that are not
models.